S4-95: bigint for transaction_id
strangely enough, transaction_id as an int works well enough on smaller (non-heavily-used) dbs... but, of course, out on long-running/busy prod servers, we NEED a bigint here (and that's what the data type is).

+++ S4 Performance/list_transactions.sql
@@ -50,7 +50,7 @@ AS
 	CREATE TABLE #core (
 		[row_number] int IDENTITY(1,1) NOT NULL,
 		[session_id] int NOT NULL,
-		[transaction_id] int NULL,
+		[transaction_id] bigint NULL,
 		[database_id] int NULL,
 		[duration] int NULL,
 		[enlisted_db_count] int NULL,
